fix: pre_build.js for windows installation (minor changes)

+++ screenpipe-app-tauri/scripts/pre_build.js
@@ -140,7 +140,18 @@ if (platform == 'linux') {
 
 /* ########## Windows ########## */
 if (platform == 'windows') {
-	const wgetPath = await findWget();
+
+    async function isInstalled(command) {
+        try {
+            await $`powershell -Command "Get-Command ${command} -ErrorAction Stop"`;
+            return true;
+        } catch {
+            console.log(`${command} not found, Please install it`)
+            return false;
+        }
+    }
+
+    const wgetPath = await findWget();
 
 	console.log('Copying screenpipe binary...');
 
@@ -171,10 +182,10 @@ if (platform == 'windows') {
 		// process.exit(1);
 	}
 
-
 	// Setup FFMPEG
-	if (!(await fs.exists(config.ffmpegRealname))) {
-		await $`${wgetPath} -nc  --no-check-certificate --show-progress ${config.windows.ffmpegUrl} -O ${config.windows.ffmpegName}.7z`
+    const is7zipInstalled = await isInstalled("7z");
+	if (!(await fs.exists(config.ffmpegRealname)) && is7zipInstalled) {
+		await $`${wgetPath} -O -nc  --no-check-certificate --show-progress ${config.windows.ffmpegUrl} -O ${config.windows.ffmpegName}.7z`
 		await $`'C:\\Program Files\\7-Zip\\7z.exe' x ${config.windows.ffmpegName}.7z`
 		await $`mv ${config.windows.ffmpegName} ${config.ffmpegRealname}`
 		await $`rm -rf ${config.windows.ffmpegName}.7z`
@@ -188,7 +199,7 @@ if (platform == 'windows') {
 
 	if (!(await fs.exists('tesseract'))) {
 		console.log('Setting up Tesseract for Windows...')
-		await $`${wgetPath} -nc  --no-check-certificate --show-progress ${tesseractUrl} -O ${tesseractInstaller}`
+		await $`${wgetPath} -O -nc  --no-check-certificate --show-progress ${tesseractUrl} -O ${tesseractInstaller}`
 		await $`"${process.cwd()}\\${tesseractInstaller}" /S /D=C:\\Program Files\\Tesseract-OCR`
 		await $`rm ${tesseractInstaller}`
 		// Replace the mv command with xcopy
@@ -204,12 +215,13 @@ if (platform == 'windows') {
 	process.env.PATH = `${process.cwd()}\\tesseract;${process.env.PATH}`
 
 	// Setup ONNX Runtime
+    const isUnzipInstalled = await isInstalled("unzip");
 	const onnxRuntimeName = "onnxruntime-win-x64-gpu-1.19.2";
 	const onnxRuntimeLibs = `${onnxRuntimeName}.zip`;
 	const onnxRuntimeUrl = `https://github.com/microsoft/onnxruntime/releases/download/v1.19.2/${onnxRuntimeLibs}`
-	if (!(await fs.exists(onnxRuntimeName))) {
+	if (!(await fs.exists(onnxRuntimeName)) && isUnzipInstalled) {
 		console.log('Setting up ONNX Runtime libraries for Windows...')
-		await $`${wgetPath} -nc  --no-check-certificate --show-progress ${onnxRuntimeUrl} -O ${onnxRuntimeLibs}`
+		await $`${wgetPath} -O -nc  --no-check-certificate --show-progress ${onnxRuntimeUrl} -O ${onnxRuntimeLibs}`
 		await $`unzip ${onnxRuntimeLibs} || tar -xf ${onnxRuntimeLibs} || echo "Done extracting"`;
 		await $`rm -rf ${onnxRuntimeLibs} || rm ${onnxRuntimeLibs} -Recurse -Force || echo "Done cleaning up zip"`;
 		console.log('ONNX Runtime libraries for Windows set up successfully.')
@@ -219,7 +231,7 @@ if (platform == 'windows') {
 
 	// Setup OpenBlas
 	if (!(await fs.exists(config.openblasRealname)) && hasFeature('openblas')) {
-		await $`${wgetPath} -nc --show-progress ${config.windows.openBlasUrl} -O ${config.windows.openBlasName}.zip`
+		await $`${wgetPath} -O -nc --show-progress ${config.windows.openBlasUrl} -O ${config.windows.openBlasName}.zip`
 		await $`"C:\\Program Files\\7-Zip\\7z.exe" x ${config.windows.openBlasName}.zip -o${config.openblasRealname}`
 		await $`rm ${config.windows.openBlasName}.zip`
 		fs.cp(path.join(config.openblasRealname, 'include'), path.join(config.openblasRealname, 'lib'), { recursive: true, force: true })
@@ -229,7 +241,7 @@ if (platform == 'windows') {
 
 	// Setup CLBlast
 	if (!(await fs.exists(config.clblastRealname)) && !hasFeature('cuda')) {
-		await $`${wgetPath} -nc --show-progress ${config.windows.clblastUrl} -O ${config.windows.clblastName}.zip`
+		await $`${wgetPath} -O -nc --show-progress ${config.windows.clblastUrl} -O ${config.windows.clblastName}.zip`
 		await $`"C:\\Program Files\\7-Zip\\7z.exe" x ${config.windows.clblastName}.zip` // 7z file inside
 		await $`"C:\\Program Files\\7-Zip\\7z.exe" x ${config.windows.clblastName}.7z` // Inner folder
 		await $`mv ${config.windows.clblastName} ${config.clblastRealname}`
@@ -237,8 +249,13 @@ if (platform == 'windows') {
 		await $`rm ${config.windows.clblastName}.7z`
 	}
 
-	// Setup vcpkg packages
-	await $`C:\\vcpkg\\vcpkg.exe install ${config.windows.vcpkgPackages}`.quiet()
+	// Setup vcpkg packages, requires VCPKG_ROOT env var
+    try {
+        await $`C:\\vcpkg\\vcpkg.exe install ${config.windows.vcpkgPackages}`
+    } catch (error) {
+        await $`vcpkg.exe install ${config.windows.vcpkgPackages}`
+    }
+    
 }
 
 async function getMostRecentBinaryPath(targetArch, paths) {
